Fix unitary eval pipeline and rerun curriculum evaluation
- Fix float32 evaluation bug: decode dataset.x and re-simulate at float64
  for comparison target; pass float32 U for model conditioning
- Fix dataset.U saved as float32 -> float64 in DatasetGenerator
- Regenerate stage1_eval with qiskit backend (was quditkit, causing
  convention mismatch with evaluation) and float64 U storage
- Fix genQC.* -> my_genQC.* in all curriculum model config.yaml files
- Change circuits_generation.py default unitary_dtype float16 -> float32
- Add reports/unitary-compilation/notes.md documenting the float32 bug
- Rerun curriculum evaluation with corrected pipeline; corrected results:
  curriculum-stage2=90.6%, baseline=96.9% on paper eval (full gate set)
  curriculum-stage2=95.3%, baseline/remote=100% on stage1_eval (Clifford)
- Update reports/curriculum-learning/notes.md and experiments/registry.yaml
  with corrected numbers and full 3-model comparison table

+++ src/quantum_diffusion/data/dataset.py
@@ -226,7 +226,7 @@ class DatasetGenerator:
                 if condition == CircuitConditionType.SRV:
                     mixed_dataset = dataset
                 elif condition == CircuitConditionType.UNITARY:
-                    dataset.U = Us.float()
+                    dataset.U = Us.double()
                     datasets_list = [dataset]
 
                     if balance_after_generation:
